Documentation fix.
